Update README.md
Problem 4.1 Trajectory Generation with Dynamical Systems [38 Points]

In this exercise we will use the Dynamic Motor Primitives (DMPs), described by the following dynamical system,

                        ¨y =
                        	

                        2 ( ( (g y) ( ˙y= )) + fw (z)) ,
                        	

                        (1)

                        z˙ =
                        	

                        z z,
                        	

                        (2)

where y is the state of the system, ˙y and ¨y are the first and second time derivatives, respectively. The attractor’s goal is denoted by g and the forcing function by fw. The parameters and control the spring-damper system. The phase variable is denoted by z and the temporal scaling coeﬃcient by . The forcing function fw is given by